Move FakeInput tests into their own test suite
